requirements.txt
fastapi
uvicorn
requests
playwright
openai
python-dotenv
pydantic